pyproject.toml
[tool.poetry]
name = "coffee-house-api-server"
version = "1.0.0"
description = ""
authors = ["Eldos <eldos.baktybekov@gmail.com>"]
readme = "README.md"
packages = [{ include = "coffee_house_api_server" }]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.9"
djangorestframework = "^3.14.0"
django-environ = "^0.11.2"
django-import-export = "^3.3.6"
httpx = "^0.26.0"
django-cors-headers = "^4.3.1"
celery = "^5.3.6"
django-celery-beat = "^2.5.0"
redis = "^5.0.1"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.7.0"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "coffee_house.settings"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]
